refactor: #2 changing the static and dynamic post page by extract Post component
