+++ src/pages/posts/dynamic/[slug].js
@@ -1,9 +1,5 @@
-import Head from "next/head";
-
 import { getAllPostSlugs, getPostDataBySlug } from "@/lib/dynamic_posts";
-import Layout from "@/components/layout/layout";
-import Date from "@/components/date";
-import utilStyles from "@/styles/utils.module.css";
+import Post from "@/components/posts/post";
 
 export async function getStaticPaths() {
   const paths = await getAllPostSlugs();
@@ -18,20 +14,6 @@ export async function getStaticProps({ params }) {
   return { props: { postData } };
 }
 
-export default function Post({ postData }) {
-  return (
-    <Layout>
-      <Head>
-        <title>{postData.title}</title>
-      </Head>
-      <article>
-        <h1 className={utilStyles.headingXl}>{postData.title}</h1>
-        <div className={utilStyles.lightText}>
-          <Date dateString={postData.createdAt} />
-        </div>
-        <br />
-        <div dangerouslySetInnerHTML={{ __html: postData.contentHtml }} />
-      </article>
-    </Layout>
-  );
-}
+const DynamicPost = ({ postData }) => <Post postData={postData} />;
+
+export default DynamicPost;
