+++ src/pages/posts/index.js
@@ -1,0 +1,3 @@
+export default function Posts() {
+  return "Hello, World";
+}
